github-mcp/requirements.txt
requests
PyGithub
PyYAML
PyJWT
